Update : Edit project box
edit homepage project box and move file HomePage.tsx to folder HomePage
